Enable license gathering from source in alpine
License gathering is a bit more problematic due to subpackages in
Alpine. There is a lot of subpackages in Alpine that all share single
codebase with their parent package, while building different subsets of
it (often with different licenses). As of now we have no way of
differentiating this thus a notice will be added alongside subpackage
support that licenses reported are for whole codebase related to the
package and for now manual verification might be required.

Signed-off-by: Alexander Mazuruk <a.mazuruk@samsung.com>